% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finaldata}
\alias{finaldata}
\title{Data to demonstrate the trimNa and trimcovNa function
Description:
  Data for 181 countries from the Uppsala Conflict Data Program and World Bank for the years 2000 to 2019.
  The dataset includes various socio-economic indicators.}
\format{
:
  A data frame with 3721 rows and 20 columns
}
\source{
: The data is sourced from the Uppsala Conflict Data Program and the World Bank (https://pubmed.ncbi.nlm.nih.gov/34582455/).
}
\usage{
finaldata
}
\description{
Data to demonstrate the trimNa and trimcovNa function
Description:
  Data for 181 countries from the Uppsala Conflict Data Program and World Bank for the years 2000 to 2019.
  The dataset includes various socio-economic indicators.
}
\examples{
  Load the data
  data("finaldata")
  str(finaldata)
 Test the trimNa function to remove countries with missing MatMort values
  finaldata1 <- trimna(finaldata, target_col = "country_name", condition_col = "MatMort")
This will remove rows from country_name column conditional on MatMort missing data
Test the trimcovNa by first defining the target column and covariate columns
target_column <- "country_name"
covariate_columns <- c("GDP", "male_edu", "temp")
Use the redefined function to clean the covariates dataset
cleaned_covariates <- trimcovNa(finaldata, target_col = target_column, covariate_cols = covariate_columns)
This will remove rows from country_name if any of the covariates have a missing value
End of Data Documentation
}
\keyword{datasets}
