% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimcovNA.R
\name{trimcovNa}
\alias{trimcovNa}
\title{Remove rows with missing values in specified covariate column}
\usage{
trimcovNa(data, target_col, covariate_cols)
}
\arguments{
\item{data}{A data frame.}

\item{target_col}{Column name where rows should be removed if there's a missing value on another specified column (covariate_cols)}

\item{covariate_cols}{Specified column names considered as covariates, if NA exists, remove that row.}
}
\value{
A data frame with rows removed.
}
\description{
This function removes rows from a column in the data frame where a specified
covariate column has a missing value.
}
\examples{
\dontrun{
# Load the NixNA package
library(trimcovna)

# Specify the target column and covariates
target_column <- "country_name"
covariate_columns <- c("GDP", "popdens", "urban", "male_edu")

# Call the function
cleaned_data <- trimcovNa(conflictdata, target_column, covariate_columns)
}

}
