% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimNA.R
\name{trimNa}
\alias{trimNa}
\title{Remove rows with missing values in a specified column}
\usage{
trimNa(data, target_col, condition_col)
}
\arguments{
\item{data}{A data frame.}

\item{target_col}{The column name where rows will be removed if condition_col has NA}

\item{condition_col}{The column name with missing values to check}
}
\value{
A data frame with rows removed.
}
\description{
This function removes rows from a data frame where a specified column
has a missing value in another specified column.
}
