% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimun.R
\name{trimun}
\alias{trimun}
\title{Remove Rows for Unique Values with Missing Covariates}
\usage{
trimun(df, unique_column, covariates)
}
\arguments{
\item{df}{A data frame.}

\item{unique_column}{A character string specifying the column with unique values.}

\item{covariates}{A character vector specifying the covariate columns.}
}
\value{
A data frame with rows removed for unique values with missing covariates.
}
\description{
The \code{trimun} function removes rows for unique values in a specified column
if at least one of the specified covariates is missing.
}
\examples{
\dontrun{
# Load the trimun package
library(trimun)

# Specify the column and covariates
unique_column <- "country_name"
covariates <- c("GDP", "popdens", "urban", "male_edu")

# Call the function
cleaned_data <- trimun(conflictdata, unique_column, covariates)
}


}
\keyword{data}
